Move dismissNotificationAlertIfNeeded method to UITestsFoundation
