+++ WordPress/WordPressUITests/Screens/BaseScreen.swift
@@ -20,19 +20,6 @@ extension BaseScreen {
     func pop() {
         navBackButton.tap()
     }
-
-    @discardableResult
-    func dismissNotificationAlertIfNeeded(_ action: FancyAlertComponent.Action = .cancel) -> Self {
-        if FancyAlertComponent.isLoaded() {
-            switch action {
-            case .accept:
-                FancyAlertComponent().acceptAlert()
-            case .cancel:
-                FancyAlertComponent().cancelAlert()
-            }
-        }
-        return self
-    }
 }
 
 private extension XCUIElementAttributes {
